// src/models/NumberTypes.ts
export enum NumberTypes {
  Byte = "UInt8",
  Int8 = "Int8",
  UInt8 = "UInt8",
  Int16BE = "Int16BE",
  Int16LE = "Int16LE",
  UInt16BE = "UInt16BE",
  UInt16LE = "UInt16LE",
  Int32BE = "Int32BE",
  Int32LE = "Int32LE",
  UInt32BE = "UInt32BE",
  UInt32LE = "UInt32LE",
  Int64BE = "Int64BE",
  Int64LE = "Int64LE",
  UInt64BE = "UInt64BE",
  UInt64LE = "UInt64LE",
  FloatBE = "FloatBE",
  FloatLE = "FloatLE",
  DoubleBE = "DoubleBE",
  DoubleLE = "DoubleLE",
}